+++ blueprints/ember-select-2/index.js
@@ -2,6 +2,6 @@ module.exports = {
   normalizeEntityName: function() {},
 
   afterInstall: function() {
-    return this.addBowerPackageToProject('select2');
+    return this.addBowerPackageToProject('select2', '3.5.2');
   }
 };