+++ tests/dummy/app/components/link-li.js
@@ -4,9 +4,9 @@ var LinkLiComponent = Ember.Component.extend({
   tagName: 'li',
   classNameBindings: ['active'],
   
-  active: function() {
+  active: Ember.computed('childViews.@each.active', function() {
     return this.get('childViews').anyBy('active');
-  }.property('childViews.@each.active')
+  })
 });
 
 export default LinkLiComponent;