+++ tests/dummy/app/controllers/examples.js
@@ -2,7 +2,7 @@ import Ember from "ember";
 
 var ExamplesController = Ember.Controller.extend({
   favouritePizza: null,
-  favouritePizzaId: Ember.computed.defaultTo('ham'),
+  favouritePizzaId: Ember.computed.oneWay('ham'),
   favouritePizzas: null,
   favouritePizzaIds: null,
   preselectedPizzaIds: ["ham", "pep", "tex"],
@@ -54,6 +54,19 @@ var ExamplesController = Ember.Controller.extend({
     }
   ],
 
+  disabledPizzas: Ember.A([
+    {
+       id: "1",
+       text: "Margherita",
+       description: "The original italian one"
+     }, {
+       id: "2",
+       text: "Cheese",
+       disabled: true,
+       description: "Can you say Laaame?"
+     }
+  ]),
+
   ingredients: [
     {
       text: "Vegetables",
@@ -102,7 +115,7 @@ var ExamplesController = Ember.Controller.extend({
       this.toggleProperty('enabled');
     },
 
-    queryPizzas: function(query, deferred) {
+    queryPizzas: function(query, deferred, isInfiniteScroll) {
       setTimeout(function() {
         var data = [];
         switch(query.term) {
@@ -119,12 +132,21 @@ var ExamplesController = Ember.Controller.extend({
                 text: 'Pizza ' + query.term + ' ' + i
               });
             }
-            deferred.resolve(data);
+            if (isInfiniteScroll) {
+              deferred.resolve({data: data, more: true});
+            } else {
+              deferred.resolve(data);
+            }
             break;
         }
       }, 300);
+    },
+
+    queryInfiniteScrollPizzas: function(query, deferred) {
+      this.send('queryPizzas', query, deferred, true);
     }
-  }
+  },
+
 });
 
 export default ExamplesController;