+++ tests/unit/components/select-2-test.js
@@ -9,7 +9,7 @@ import startApp from "../../helpers/start-app";
 /*
   Test Fixtures
  */
-var simpleContent = [
+var simpleContent = Ember.A([
   {
     id: true,
     text: "Margherita"
@@ -22,9 +22,13 @@ var simpleContent = [
   }, {
     id: "haw",
     text: "Hawaii"
+  }, {
+    id: "lame",
+    disabled: true,
+    text: "Plain Cheese"
   }
-];
-var additionalContent = [
+]);
+var additionalContent = Ember.A([
   {
     id: "cal",
     text: "Calzone"
@@ -32,7 +36,7 @@ var additionalContent = [
     id: "bbq",
     text: "Barbecue"
   }
-];
+]);
 
 
 var App, component;
@@ -49,44 +53,44 @@ moduleForComponent('select-2', 'Select2Component', {
 });
 
 
-test("it renders", function() {
-  expect(2);
+test("it renders", function(assert) {
+  assert.expect(2);
 
-  equal(component._state, 'preRender');
+  assert.equal(component._state, 'preRender');
 
   // appends the component to the page
-  this.append();
-  equal(component._state, 'inDOM');
+  this.render();
+  assert.equal(component._state, 'inDOM');
 });
 
 
-test("it initializes select2 plugin", function() {
-  expect(2);
+test("it initializes select2 plugin", function(assert) {
+  assert.expect(2);
 
   // append the component to the DOM
-  this.append();
+  this.render();
 
-  ok(component.$().data('select2'), "has select2 data attribute");
+  assert.ok(component.$().data('select2'), "has select2 data attribute");
 
-  ok($(".select2-container").length, "inserts container into dom");
+  assert.ok($(".select2-container").length, "inserts container into dom");
 });
 
 
-test("it supports placeholder text", function() {
+test("it supports placeholder text", function(assert) {
   var placeholder = "unit testing rocks";
 
   component.set('placeholder', placeholder);
 
-  this.append();
+  this.render();
 
-  equal($('.select2-chosen').text(), placeholder, "has placeholder text");
+  assert.equal($('.select2-chosen').text(), placeholder, "has placeholder text");
 });
 
 
-test("it shows options when opened", function() {
-  expect(2);
+test("it shows options when opened", function(assert) {
+  assert.expect(2);
 
-  this.append();
+  this.render();
 
   component.set('content', simpleContent);
 
@@ -94,16 +98,16 @@ test("it shows options when opened", function() {
   click('.select2-choice');
 
   andThen(function() {
-    equal($('.select2-results li').length, simpleContent.length, "has correct options length");
-    equal($('.select2-results li').text(), simpleContent.getEach('text').join(''), "display correct text");
+    assert.equal($('.select2-results li').length, simpleContent.length, "has correct options length");
+    assert.equal($('.select2-results li').text(), simpleContent.getEach('text').join(''), "display correct text");
   });
 });
 
 
-test("it sets value to selected object in single selection mode", function() {
-  expect(4);
+test("it sets value to selected object in single selection mode", function(assert) {
+  assert.expect(4);
 
-  this.append();
+  this.render();
 
   component.set('content', simpleContent);
 
@@ -113,22 +117,22 @@ test("it sets value to selected object in single selection mode", function() {
   click('.select2-results li:nth-child(3)', 'body');
 
   andThen(function() {
-    strictEqual(component.get('value'), simpleContent[2], "selects correct item");
-    equal($('.select2-chosen').text(), simpleContent[2].text, "has correct text");
+    assert.strictEqual(component.get('value'), simpleContent[2], "selects correct item");
+    assert.equal($('.select2-chosen').text(), simpleContent[2].text, "has correct text");
 
     // select another option just to make sure
     click('.select2-choice');
     click('.select2-results li:nth-child(1)', 'body');
 
     andThen(function() {
-      strictEqual(component.get('value'), simpleContent[0], "selects correct item");
-      equal($('.select2-chosen').text(), simpleContent[0].text, "has correct text");
+      assert.strictEqual(component.get('value'), simpleContent[0], "selects correct item");
+      assert.equal($('.select2-chosen').text(), simpleContent[0].text, "has correct text");
     });
   });
 });
 
-test("it sends `didSelect` action once when selection has been changed", function() {
-  expect(1);
+test("it sends `didSelect` action once when selection has been changed", function(assert) {
+  assert.expect(1);
 
   var controller = {
     selectionChanged: function() {}
@@ -141,7 +145,7 @@ test("it sends `didSelect` action once when selection has been changed", functio
     didSelect: 'selectionChanged'
   });
 
-  this.append();
+  this.render();
 
   // open options by clicking on the element
   click('.select2-choice');
@@ -149,7 +153,7 @@ test("it sends `didSelect` action once when selection has been changed", functio
   click('.select2-results li:nth-child(3)', 'body');
 
   andThen(function() {
-    ok(spy.calledOnce, "callback after selection changed has been executed");
+    assert.ok(spy.calledOnce, "callback after selection changed has been executed");
   });
 });
 
@@ -157,14 +161,14 @@ test("it sends `didSelect` action once when selection has been changed", functio
   This test assures, that when sending the `didSelect` action, bindings have
   already been synced. Test for PR #75
  */
-test("component has correct value before `didSelect` fires", function() {
-  expect(1);
+test("component has correct value before `didSelect` fires", function(assert) {
+  assert.expect(1);
 
   var controller = Ember.Object.create({
     component: component,
     valueBinding: 'component.value',
     selectionChanged: function() {
-      strictEqual(this.get('value'), simpleContent[2], "selects correct item");
+      assert.strictEqual(this.get('value'), simpleContent[2], "selects correct item");
     }
   });
 
@@ -174,7 +178,7 @@ test("component has correct value before `didSelect` fires", function() {
     didSelect: 'selectionChanged'
   });
 
-  this.append();
+  this.render();
 
   // open options by clicking on the element
   click('.select2-choice');
@@ -182,13 +186,70 @@ test("component has correct value before `didSelect` fires", function() {
   click('.select2-results li:nth-child(3)', 'body');
 });
 
-test("it supports the allowClear option", function() {
-  expect(3);
+
+test("it sends `didSelect` action having the new value as the first argument", function(assert) {
+  assert.expect(2);
+
+  var controller = {
+    selectionChanged: function() {}
+  };
+  var spy = sinon.spy(controller, 'selectionChanged');
+
+  component.setProperties({
+    targetObject: controller,
+    content: simpleContent,
+    didSelect: 'selectionChanged'
+  });
+
+  this.render();
+
+  // open options by clicking on the element
+  click('.select2-choice');
+  // then select an option
+  click('.select2-results li:nth-child(3)', 'body');
+
+  andThen(function() {
+    assert.ok(spy.calledOnce, "callback after selection changed has been executed");
+    assert.ok(spy.calledWith(simpleContent[2]), "has new value as argument");
+  });
+});
+
+
+test("it sends `didSelect` action having a reference to itself as the second argument", function(assert) {
+  assert.expect(2);
+
+  var controller = {
+    selectionChanged: function() {}
+  };
+  var spy = sinon.spy(controller, 'selectionChanged');
+
+  component.setProperties({
+    targetObject: controller,
+    content: simpleContent,
+    didSelect: 'selectionChanged'
+  });
+
+  this.render();
+
+  // open options by clicking on the element
+  click('.select2-choice');
+  // then select an option
+  click('.select2-results li:nth-child(3)', 'body');
+
+  andThen(function() {
+    assert.ok(spy.calledOnce, "callback after selection changed has been executed");
+    assert.ok(spy.calledWith(simpleContent[2], component), "has reference to self as argument");
+  });
+});
+
+
+test("it supports the allowClear option", function(assert) {
+  assert.expect(3);
 
   component.set('placeholder', 'Select a value'); // placeholder is required for allowClear
   component.set('allowClear', true);
 
-  this.append();
+  this.render();
 
   component.set('content', simpleContent);
 
@@ -198,35 +259,35 @@ test("it supports the allowClear option", function() {
   click('.select2-results li:nth-child(3)', 'body');
 
   andThen(function() {
-    strictEqual(component.get('value'), simpleContent[2], "selects correct item");
-    equal($('.select2-chosen').text(), simpleContent[2].text, "has correct text");
+    assert.strictEqual(component.get('value'), simpleContent[2], "selects correct item");
+    assert.equal($('.select2-chosen').text(), simpleContent[2].text, "has correct text");
 
     // Click the remove option x
     click('.select2-search-choice-close:visible');
 
     andThen(function() {
-      strictEqual(component.get('value'), null, "unselects the selected item");
+      assert.strictEqual(component.get('value'), null, "unselects the selected item");
     });
   });
 });
 
-test("it alerts if allowClear is set without a placeholder", function() {
-  expect(1);
+test("it alerts if allowClear is set without a placeholder", function(assert) {
+  assert.expect(1);
 
   component.set('placeholder', undefined);
   component.set('allowClear', true);
 
   try {
-    this.append();
+    this.render();
   } catch (e) {
-    equal(e.message, 'Assertion Failed: To use allowClear, you have to specify a placeholder', 'throws and error');
+    assert.equal(e.message, 'Assertion Failed: To use allowClear, you have to specify a placeholder', 'throws and error');
   }
 });
 
-test("it sets value to selected object's optionValuePath in single selection mode", function() {
-  expect(4);
+test("it sets value to selected object's optionValuePath in single selection mode", function(assert) {
+  assert.expect(4);
 
-  this.append();
+  this.render();
 
   component.set('content', simpleContent);
   component.set('optionValuePath', 'id');
@@ -237,137 +298,164 @@ test("it sets value to selected object's optionValuePath in single selection mod
   click('.select2-results li:nth-child(3)', 'body');
 
   andThen(function() {
-    strictEqual(component.get('value'), simpleContent[2].id, "selects correct item");
-    equal($('.select2-chosen').text(), simpleContent[2].text, "has correct text");
+    assert.strictEqual(component.get('value'), simpleContent[2].id, "selects correct item");
+    assert.equal($('.select2-chosen').text(), simpleContent[2].text, "has correct text");
 
     // select another option just to make sure
     click('.select2-choice');
     click('.select2-results li:nth-child(1)', 'body');
 
     andThen(function() {
-      strictEqual(component.get('value'), simpleContent[0].id, "selects correct item");
-      equal($('.select2-chosen').text(), simpleContent[0].text, "has correct text");
+      assert.strictEqual(component.get('value'), simpleContent[0].id, "selects correct item");
+      assert.equal($('.select2-chosen').text(), simpleContent[0].text, "has correct text");
     });
   });
 });
 
 
-test("it reacts to external value change", function() {
-  expect(2);
+test("it allows individual options to be disabled", function(assert) {
+  assert.expect(3);
 
-  this.append();
+  this.render();
+
+  component.set('content', simpleContent);
+  component.set('optionValuePath', 'id');
+
+  //select an enabled option
+  click('.select2-choice');
+  click('.select2-results li:nth-child(2)', 'body');
+
+  //click disabled option
+  click('.select2-choice');
+  click('.select2-results li:nth-child(5)', 'body');
+
+  andThen(function() {
+    //selection should remain unchanged
+    assert.strictEqual(component.get('value'), simpleContent[1].id, "selection remains unchanged");
+    assert.equal($('.select2-chosen').text(), simpleContent[1].text, "remains unchanged");
+
+    //should have disabled styling
+    assert.equal($('.select2-results li:nth-child(5)').hasClass('select2-disabled'), true, "has disabled class");
+  });
+
+});
+
+test("it reacts to external value change", function(assert) {
+  assert.expect(2);
+
+  this.render();
 
   component.set('content', simpleContent);
 
-  equal($('.select2-chosen').text(), '', "has empty selection text on start");
+  assert.equal($('.select2-chosen').text(), '', "has empty selection text on start");
 
   component.set('value', simpleContent[1]);
 
-  equal($('.select2-chosen').text(), simpleContent[1].text, "has correct selection text after value change");
+  assert.equal($('.select2-chosen').text(), simpleContent[1].text, "has correct selection text after value change");
 });
 
 
-test("it reacts to external value change with optionValuePath", function() {
-  expect(2);
+test("it reacts to external value change with optionValuePath", function(assert) {
+  assert.expect(2);
 
-  this.append();
+  this.render();
 
   component.set('content', simpleContent);
   component.set('optionValuePath', 'id');
 
-  equal($('.select2-chosen').text(), '', "has empty selection text on start");
+  assert.equal($('.select2-chosen').text(), '', "has empty selection text on start");
 
   component.set('value', simpleContent[1].id);
 
-  equal($('.select2-chosen').text(), simpleContent[1].text, "has correct selection text after value change");
+  assert.equal($('.select2-chosen').text(), simpleContent[1].text, "has correct selection text after value change");
 });
 
 
-test("it sets value to array of selected objects in multiple selection mode", function() {
-  expect(7);
+test("it sets value to array of selected objects in multiple selection mode", function(assert) {
+  assert.expect(7);
 
   component.set('multiple', true);
 
-  this.append();
+  this.render();
 
   component.set('content', simpleContent);
 
-  equal($('.select2-choices').text().trim(), '', "has empty selection text on start");
+  assert.equal($('.select2-choices').text().trim(), '', "has empty selection text on start");
 
   // select an option after opening the dropdown
   click('.select2-choices');
   click('.select2-results li:nth-child(3)', 'body');
 
   andThen(function() {
-    deepEqual(component.get('value'), [simpleContent[2]], "has correct value");
-    equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[2].text, "displays correct text");
+    assert.deepEqual(component.get('value'), [simpleContent[2]], "has correct value");
+    assert.equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[2].text, "displays correct text");
 
     // select another item
     click('.select2-choices');
     click('.select2-results li:nth-child(1)', 'body');
 
     andThen(function() {
-      deepEqual(component.get('value'), [simpleContent[2], simpleContent[0]], "has correct value");
-      equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[2].text + simpleContent[0].text, "displays correct text");
+      assert.deepEqual(component.get('value'), [simpleContent[2], simpleContent[0]], "has correct value");
+      assert.equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[2].text + simpleContent[0].text, "displays correct text");
 
       // remove the first item again
       click('.select2-search-choice:nth-child(1) .select2-search-choice-close');
 
       andThen(function() {
-        deepEqual(component.get('value'), [simpleContent[0]], "has correct value");
-        equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[0].text, "displays correct text");
+        assert.deepEqual(component.get('value'), [simpleContent[0]], "has correct value");
+        assert.equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[0].text, "displays correct text");
       });
     });
   });
 });
 
 
-test("it sets value to array of selected objects' optionValuePaths in multiple selection mode", function() {
-  expect(7);
+test("it sets value to array of selected objects' optionValuePaths in multiple selection mode", function(assert) {
+  assert.expect(7);
 
   component.set('multiple', true);
 
-  this.append();
+  this.render();
 
   component.set('content', simpleContent);
   component.set('optionValuePath', 'id');
 
-  equal($('.select2-choices').text().trim(), '', "has empty selection text on start");
+  assert.equal($('.select2-choices').text().trim(), '', "has empty selection text on start");
 
   // select an option after opening the dropdown
   click('.select2-choices');
   click('.select2-results li:nth-child(3)', 'body');
 
   andThen(function() {
-    deepEqual(component.get('value'), [simpleContent[2].id], "has correct value");
-    equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[2].text, "displays correct text");
+    assert.deepEqual(component.get('value'), [simpleContent[2].id], "has correct value");
+    assert.equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[2].text, "displays correct text");
 
     // select another item
     click('.select2-choices');
     click('.select2-results li:nth-child(1)', 'body');
 
     andThen(function() {
-      deepEqual(component.get('value'), [simpleContent[2].id, simpleContent[0].id], "has correct value");
-      equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[2].text + simpleContent[0].text, "displays correct text");
+      assert.deepEqual(component.get('value'), [simpleContent[2].id, simpleContent[0].id], "has correct value");
+      assert.equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[2].text + simpleContent[0].text, "displays correct text");
 
       // remove the first item again
       click('.select2-search-choice:nth-child(1) .select2-search-choice-close');
 
       andThen(function() {
-        deepEqual(component.get('value'), [simpleContent[0].id], "has correct value");
-        equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[0].text, "displays correct text");
+        assert.deepEqual(component.get('value'), [simpleContent[0].id], "has correct value");
+        assert.equal($('.select2-choices').text().replace(/ /g, ''), simpleContent[0].text, "displays correct text");
       });
     });
   });
 });
 
 
-test("it reacts to changing the content array", function() {
-  expect(4);
+test("it reacts to changing the content array", function(assert) {
+  assert.expect(4);
 
-  this.append();
+  this.render();
 
-  var content = [];
+  var content = Ember.A();
   content.pushObjects(simpleContent);
 
   component.set('content', content);
@@ -376,8 +464,8 @@ test("it reacts to changing the content array", function() {
   click('.select2-choice');
 
   andThen(function() {
-    equal($('.select2-results li').length, simpleContent.length, "has correct options length");
-    equal($('.select2-results li').text(), simpleContent.getEach('text').join(''), "display correct text");
+    assert.equal($('.select2-results li').length, simpleContent.length, "has correct options length");
+    assert.equal($('.select2-results li').text(), simpleContent.getEach('text').join(''), "display correct text");
 
     click('.select2-choice');
 
@@ -386,19 +474,19 @@ test("it reacts to changing the content array", function() {
     click('.select2-choice');
 
     andThen(function() {
-      equal($('.select2-results li').length, simpleContent.length + additionalContent.length, "has correct options length");
-      equal($('.select2-results li').text(), simpleContent.getEach('text').join('') + additionalContent.getEach('text').join(''), "display correct text");
+      assert.equal($('.select2-results li').length, simpleContent.length + additionalContent.length, "has correct options length");
+      assert.equal($('.select2-results li').text(), simpleContent.getEach('text').join('') + additionalContent.getEach('text').join(''), "display correct text");
     });
   });
 });
 
 
-test("it is disabled when its selection contains values not in the content array", function() {
-  expect(2);
+test("it is disabled when its selection contains values not in the content array", function(assert) {
+  assert.expect(2);
 
-  this.append();
+  this.render();
 
-  var content = [];
+  var content = Ember.A();
   content.pushObjects(simpleContent);
 
   component.set('content', content);
@@ -406,25 +494,26 @@ test("it is disabled when its selection contains values not in the content array
   component.set('optionValuePath', 'id');
   component.set('value', ['bbq']);
 
-  ok($('.select2-container').hasClass('select2-container-disabled'), "is disabled");
+  assert.ok($('.select2-container').hasClass('select2-container-disabled'), "is disabled");
 
   content.pushObjects(additionalContent);
 
-  ok(!$('.select2-container').hasClass('select2-container-disabled'), "is enabled");
+  assert.ok(!$('.select2-container').hasClass('select2-container-disabled'), "is enabled");
 });
 
-test("Change event does not trigger an autorun", function() {
-  expect(1);
 
-  this.append();
-  var content = [];
+test("Change event does not trigger an autorun", function(assert) {
+  assert.expect(1);
+
+  this.render();
+  var content = Ember.A();
   content.pushObjects(simpleContent);
   component.set('content', content);
   component.set('optionValuePath', 'id');
 
   Ember.addObserver(component, 'value', function() {
     Ember.run.schedule('afterRender', function() {
-      ok(true);
+      assert.ok(true);
     });
   });