+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # ember-select-2 changelog
 
-### TBD
+### 1.3.0 (May 5, 2015)
 
+* Add ability for infinite scroll (#85, thanks, cooperjbrandon!)
 * The `didSelect` action should wait for databinding (#75, thanks, pierrickrouxel!)
 * Make select2 assets optional (#71, thanks, minichate!)
+* The `didSelect` action passes new value and reference to component (suggested in #76, thanks, DENIELER!)
+* Add ability to set a custom separator for values (#88, thanks drapergeek!)
+* Fix a crash with disabled prototype extension (#79, thanks, huafu!)
+* Add the ability to pass a `tabindex` (suggested in #53, thanks, ryanjm!)
+* Add support for disabled options (#47, thanks, kevintraver!)
+* Add `optionSelectedLabelPath` option (#66, thanks, ziggythehamster!)
+* Fix possible incopatibility with disabled prototype extension
 
 ### 1.2.0 (February 11, 2015)